Same speed but all on the grid
I can see them all on the grid but I'm having issues when specifying each car's speed. I think there's a problem with the movement when I set the new direction of x and y. Now it's fine but only when all are on the same speed. I also need to fix the flickering lights on the circles.